+++ src/lib/lua_mympd_state.h
@@ -20,13 +20,13 @@ enum lua_mympd_state_type {
 
 struct t_lua_mympd_state_value {
     sds p;
-    long i;
+    long long i;
     double f;
     bool b;
 };
 
 void lua_mympd_state_set_p(struct t_list *lua_mympd_state, const char *k, const char *v);
-void lua_mympd_state_set_i(struct t_list *lua_mympd_state, const char *k, long v);
+void lua_mympd_state_set_i(struct t_list *lua_mympd_state, const char *k, long long v);
 void lua_mympd_state_set_u(struct t_list *lua_mympd_state, const char *k, unsigned v);
 void lua_mympd_state_set_b(struct t_list *lua_mympd_state, const char *k, bool v);
 void lua_mympd_state_free(struct t_list *lua_mympd_state);
